refactor: 테스트 도메인 Fixture 분리
