+++ src/test/java/wooteco/subway/acceptanceTest/section/SectionAcceptanceTest.java
@@ -1,21 +1,21 @@
 package wooteco.subway.acceptanceTest.section;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.DEFAULT_SECTION_DISTANCE;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.LINE_ID;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.NEW_SECTION_DISTANCE;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.NEW_STATION;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.NEW_STATION_ID;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.NEW_STATION_NAME;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.STATION_1;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.STATION_2;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.STATION_3;
-import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.STATION_4;
 import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.assertSectionCreatedResponseBody;
 import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.assertStationsList;
 import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.createLineWithSectionsOf;
 import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.createSection;
 import static wooteco.subway.acceptanceTest.section.SectionAcceptanceTestUtils.deleteSection;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.DEFAULT_SECTION_DISTANCE;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.LINE_ID;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.NEW_SECTION_DISTANCE;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.NEW_STATION;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.NEW_STATION_ID;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.NEW_STATION_NAME;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.STATION_1;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.STATION_2;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.STATION_3;
+import static wooteco.subway.acceptanceTest.fixture.DomainFixtures.STATION_4;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
